kinda did a lot of shit idk read the description
enhanced the login system (fixed bugs, made it accessible in all intended pages). 

Added register and login with regular email and password.

HasalaTech is going mobile let’s goooooooo (lol). It’s kind of a long story actually you remember how the sign in and sign up buttons didn’t appear on mobile devices? Yeah that was kinda my fault, apparently it was supposed to be hidden and toggled by this burger menu but for some reason i deleted it. Anyway i put it back now and it’s working perfectly.